+++ scripts/stageapi/library/streaks.lua
@@ -1,279 +1,6 @@
 local shared = require("scripts.stageapi.shared")
 local mod = require("scripts.stageapi.mod")
 
-StageAPI.LogMinor("Loading Core Functions")
-
-StageAPI.RandomRNG = RNG()
-StageAPI.RandomRNG:SetSeed(Random(), 0)
-function StageAPI.Random(a, b, rng)
-    rng = rng or StageAPI.RandomRNG
-    if a and b then
-        -- TODO remove after Rev update
-        if b - a < 0 then
-            StageAPI.LogErr('Bad Random Range! ' .. a .. ', ' .. b)
-            return b - a
-        end
-        return rng:Next() % (b - a + 1) + a
-    elseif a then
-        -- TODO remove after Rev update
-        if a < 0 then
-            StageAPI.LogErr('Bad Random Max! ' .. a)
-            return a
-        end
-        return rng:Next() % (a + 1)
-    end
-    return rng:Next()
-end
-
-function StageAPI.RandomFloat(a, b, rng)
-    rng = rng or StageAPI.RandomRNG
-    local rand = rng:RandomFloat()
-    if a and b then
-        return (rand * (b - a)) + a
-    elseif a then
-        return rand * a
-    end
-
-    return rand
-end
-
-function StageAPI.WeightedRNG(args, rng, key, preCalculatedWeight, floatWeights) -- takes tables {{obj, weight}, {"pie", 3}, {555, 0}}
-    local weight_value = preCalculatedWeight or 0
-    local iterated_weight = 1
-    if not preCalculatedWeight then
-        for _, potentialObject in ipairs(args) do
-            if key then
-                weight_value = weight_value + potentialObject[key]
-            else
-                weight_value = weight_value + potentialObject[2]
-            end
-
-            if weight_value % 1 ~= 0 then -- if any weight is a float, use float RNG
-                floatWeights = true
-            end
-        end
-    end
-
-    rng = rng or StageAPI.RandomRNG
-    local random_chance
-    if weight_value % 1 == 0 and not floatWeights then
-        random_chance = StageAPI.Random(1, weight_value, rng)
-    else
-        random_chance = StageAPI.RandomFloat(1, weight_value, rng)
-    end
-
-    for i, potentialObject in ipairs(args) do
-        if key then
-            iterated_weight = iterated_weight + potentialObject[key]
-        else
-            iterated_weight = iterated_weight + potentialObject[2]
-        end
-
-        if iterated_weight > random_chance then
-            local ret = potentialObject
-            if key then
-                return ret, i
-            else
-                return ret[1], i
-            end
-        end
-    end
-end
-
-StageAPI.Class = {}
-function StageAPI.ClassInit(tbl, ...)
-    local inst = {}
-    setmetatable(inst, tbl)
-    tbl.__index = tbl
-    tbl.__call = StageAPI.ClassInit
-
-    if inst.AllowMultipleInit or not inst.Initialized then
-        inst.Initialized = true
-        if inst.Init then
-            inst:Init(...)
-        end
-
-        if inst.PostInit then
-            inst:PostInit(...)
-        end
-    else
-        if inst.InheritInit then
-            inst:InheritInit(...)
-        end
-    end
-
-    return inst
-end
-
-function StageAPI.Class:Init(Type, AllowMultipleInit)
-    self.Type = Type
-    self.AllowMultipleInit = AllowMultipleInit
-    self.Initialized = false
-end
-
-setmetatable(StageAPI.Class, {
-    __call = StageAPI.ClassInit
-})
-
-StageAPI.Callbacks = {}
-
-local function Reverse_Iterator(t,i)
-    i=i-1
-    local v=t[i]
-    if v==nil then return v end
-    return i,v
-end
-
-function StageAPI.ReverseIterate(t)
-    return Reverse_Iterator, t, #t+1
-end
-
-function StageAPI.AddCallback(modID, id, priority, fn, ...)
-    if not StageAPI.Callbacks[id] then
-        StageAPI.Callbacks[id] = {}
-    end
-
-    local index = 1
-
-    for i, callback in StageAPI.ReverseIterate(StageAPI.Callbacks[id]) do
-        if priority >= callback.Priority then
-            index = i + 1
-            break
-        end
-    end
-
-    table.insert(StageAPI.Callbacks[id], index, {
-        Priority = priority,
-        Function = fn,
-        ModID = modID,
-        Params = {...}
-    })
-end
-
-function StageAPI.UnregisterCallbacks(modID)
-    for id, callbacks in pairs(StageAPI.Callbacks) do
-        for i, callback in StageAPI.ReverseIterate(callbacks) do
-            if callback.ModID == modID then
-                table.remove(callbacks, i)
-            end
-        end
-    end
-end
-
-StageAPI.UnregisterCallbacks("StageAPI")
-
-function StageAPI.GetCallbacks(id)
-    return StageAPI.Callbacks[id] or {}
-end
-
-function StageAPI.CallCallbacks(id, breakOnFirstReturn, ...)
-    for _, callback in ipairs(StageAPI.GetCallbacks(id)) do
-        local ret = callback.Function(...)
-        if breakOnFirstReturn and ret ~= nil then
-            return ret
-        end
-    end
-end
-
-function StageAPI.IsIn(tbl, v, fn)
-    fn = fn or ipairs
-    for k, v2 in fn(tbl) do
-        if v2 == v then
-            return k or true
-        end
-    end
-end
-
-function StageAPI.Copy(tbl)
-    local t = {}
-    for k, v in pairs(tbl) do
-        t[k] = v
-    end
-    return t
-end
-
-function StageAPI.DeepCopy(tbl)
-    local t = {}
-    for k, v in pairs(tbl) do
-        if type(v) == "table" then
-            t[k] = StageAPI.DeepCopy(v)
-        else
-            t[k] = v
-        end
-    end
-
-    return t
-end
-
-function StageAPI.Merged(...)
-    local t = {}
-    for _, tbl in ipairs({...}) do
-        local orderedIndices = {}
-        for i, v in ipairs(tbl) do
-            orderedIndices[i] = true
-            t[#t + 1] = v
-        end
-
-        for k, v in pairs(tbl) do
-            if not orderedIndices[k] then
-                t[k] = v
-            end
-        end
-    end
-
-    return t
-end
-
-function StageAPI.GetPlayingAnimation(sprite, animations)
-    for _, anim in ipairs(animations) do
-        if sprite:IsPlaying(anim) then
-            return anim
-        end
-    end
-end
-
-function StageAPI.VectorToGrid(x, y, width)
-    width = width or shared.Room:GetGridWidth()
-    return width + 1 + (x + width * y)
-end
-
-function StageAPI.GridToVector(index, width)
-    width = width or shared.Room:GetGridWidth()
-    return (index % width) - 1, (math.floor(index / width)) - 1
-end
-
-function StageAPI.GetScreenBottomRight()
-    return shared.Room:GetRenderSurfaceTopLeft() * 2 + Vector(442,286)
-end
-
-function StageAPI.GetScreenCenterPosition()
-    return StageAPI.GetScreenBottomRight() / 2
-end
-
-StageAPI.DefaultScreenSize = Vector(480, 270)
-function StageAPI.GetScreenScale(vec)
-    local bottomRight = StageAPI.GetScreenBottomRight()
-    if vec then
-        return Vector(bottomRight.X / StageAPI.DefaultScreenSize.X, bottomRight.Y / StageAPI.DefaultScreenSize.Y)
-    else
-        return bottomRight.X / StageAPI.DefaultScreenSize.X, bottomRight.Y / StageAPI.DefaultScreenSize.Y
-    end
-end
-
-function StageAPI.Lerp(first, second, percent)
-    return first * (1 - percent) + second * percent
-end
-
-function StageAPI.FillBits(count)
-    return (1 << count) - 1
-end
-
-function StageAPI.GetBits(bits, startBit, count)
-    bits = bits >> startBit
-    bits = bits & StageAPI.FillBits(count)
-    return bits
-end
-
 local TextStreakScales = {
     [0] = Vector(3,0.2),     [1] = Vector(2.6,0.36),
     [2] = Vector(2.2,0.52),  [3] = Vector(1.8,0.68),
@@ -475,20 +202,3 @@ end
 
 mod:AddCallback(ModCallbacks.MC_POST_UPDATE, StageAPI.UpdateTextStreak)
 mod:AddCallback(ModCallbacks.MC_POST_RENDER, StageAPI.RenderTextStreak)
-
-
-function StageAPI.SpawnFloorEffect(pos, velocity, spawner, anm2, loadGraphics, variant)
-    local creep = StageAPI.E.FloorEffectCreep
-    local eff = Isaac.Spawn(creep.T, creep.V, creep.S, pos or Vector.Zero, velocity or Vector.Zero, spawner)
-    eff.Variant = variant or StageAPI.E.FloorEffect.V
-
-    if anm2 then
-        eff:GetSprite():Load(anm2, loadGraphics)
-    end
-
-    return eff
-end
-
-function StageAPI.InStartingRoom()
-    return shared.Level:GetCurrentRoomDesc().SafeGridIndex == shared.Level:GetStartingRoomIndex()
-end
