+++ scripts/stageapi/loadOrder.lua
@@ -2,7 +2,7 @@ return {
     "scripts.stageapi.basic",
     "scripts.stageapi.basiclibrary",
     "scripts.stageapi.overlays",
-    "scripts.stageapi.roomslist",
+    "scripts.stageapi.rooms",
     "scripts.stageapi.customgrids",
     "scripts.stageapi.extrarooms",
     "scripts.stageapi.customdoors",
@@ -21,7 +21,7 @@ return {
     "scripts.stageapi.waves",
     "scripts.stageapi.customFloorGen",
     "scripts.stageapi.basementRenovator",
-    "scripts.stageapi.data",
-    "scripts.stageapi.modcompat.minimapAPI",
-    "scripts.stageapi.modcompat.changelogs",
+    "scripts.stageapi.reimpl",
+    "scripts.stageapi.modcompat",
+    "scripts.stageapi.changelogs",
 }
