+++ scripts/stageapi/reimpl.lua
@@ -1,5 +1,8 @@
 StageAPI.LogMinor("Loading Reimplementation Data")
 
+-- originally was data.lua, used more
+-- descriptive name
+
 local subModules = {
     "doors",
     "gfx",
@@ -7,8 +10,9 @@ local subModules = {
     -- "bossAnimations", not needed as it was empty in the original version
     "bosses",
     "floorInfo",
+    "music",
 }
 
 for _, subModule in ipairs(subModules) do
-    include("scripts.stageapi.data." .. subModule)
+    include("scripts.stageapi.reimpl." .. subModule)
 end
