+++ scripts/stageapi/reimpl/bosses.lua
