+++ scripts/stageapi/reimpl/doors.lua
